requirements.txt
streamlit
python-dotenv
openai
langchain
langchain_openai
langchain_community
chromadb
PyMuPDF
pytesseract
Pillow
